Added Hoster class, old hoster renamed to VpsHoster
